requirements.txt
numpy
opencv-python
smbus
RPi.GPIO